+++ Sources/Dynamic Price/GoogleAdManager.swift
@@ -5,19 +5,16 @@
 //  Created by Jason Sznol on 5/27/23.
 //
 
-import NimbusKit
 import GoogleMobileAds
-import UIKit
-
+import NimbusKit
 #if canImport(NimbusSDK) // CocoaPods
 import NimbusSDK
 #else                    // Swift Package Manager
 import NimbusGAMKit
 #endif
+import UIKit
 
-private extension Bundle {
-    static let gamPlacementId = Bundle.main.infoDictionary?["Google Placement ID"] as? String ?? ""
-}
+fileprivate let gamPlacementId = Bundle.main.infoDictionary?["Google Placement ID"] as? String ?? ""
 
 final class GoogleDynamicPriceViewController: DemoViewController {
     
@@ -41,9 +38,8 @@ final class GoogleDynamicPriceViewController: DemoViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
         
-        if Bundle.gamPlacementId.isEmpty {
-            showCustomAlert("google_placement_id")
-        }
+        if gamPlacementId.isEmpty { showCustomAlert("google_placement_id") }
+        
         setupAdRendering()
     }
     
@@ -53,7 +49,7 @@ final class GoogleDynamicPriceViewController: DemoViewController {
             bannerView = GAMBannerView(adSize: GADAdSizeBanner)
             guard let bannerView else { return }
             bannerView.rootViewController = self
-            bannerView.adUnitID = Bundle.gamPlacementId
+            bannerView.adUnitID = gamPlacementId
             bannerView.delegate = self
             bannerView.accessibilityIdentifier = "google_ad_view"
             
@@ -85,7 +81,7 @@ final class GoogleDynamicPriceViewController: DemoViewController {
             bannerView = GAMBannerView(adSize: GADAdSizeMediumRectangle)
             guard let bannerView else { return }
             bannerView.rootViewController = self
-            bannerView.adUnitID = Bundle.gamPlacementId
+            bannerView.adUnitID = gamPlacementId
             bannerView.delegate = self
             bannerView.validAdSizes = [NSValueFromGADAdSize(GADAdSizeFromCGSize(CGSize(width: 400, height: 300)))]
             bannerView.accessibilityIdentifier = "google_ad_view"
@@ -123,7 +119,7 @@ final class GoogleDynamicPriceViewController: DemoViewController {
             bannerView = GAMBannerView(adSize: GADAdSizeMediumRectangle)
             guard let bannerView else { return }
             bannerView.rootViewController = self
-            bannerView.adUnitID = Bundle.gamPlacementId
+            bannerView.adUnitID = gamPlacementId
             bannerView.delegate = self
             bannerView.validAdSizes = [NSValueFromGADAdSize(GADAdSizeFromCGSize(CGSize(width: 400, height: 300)))]
             bannerView.accessibilityIdentifier = "google_ad_view"
@@ -230,7 +226,7 @@ extension GoogleDynamicPriceViewController: NimbusRequestManagerDelegate {
         if adType != .dynamicPriceInterstitial{
             bannerView?.load(gamRequest)
         } else {
-            GADInterstitialAd.load(withAdUnitID: Bundle.gamPlacementId, request: gamRequest) { gadAd, error in
+            GADInterstitialAd.load(withAdUnitID: gamPlacementId, request: gamRequest) { gadAd, error in
                 if let error {
                     print("Failed to load dynamic price interstitial ad with error: \(error.localizedDescription)")
                     return
