--- Sources/UI/UIColor+Extension.swift
@@ -1,14 +1,0 @@
-//
-//  UIColor+Extension.swift
-//  nimbus-ios-sample
-//
-//  Created by Victor Takai on 08/11/21.
-//
-
-import UIKit
-
-extension UIColor {
-    static var pink = UIColor(named: "pink")!
-    static var raisingBlack = UIColor(named: "raisinBlack")!
-    static var turquoise = UIColor(named: "turquoise")!
-}
